requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
Django==5.1.3
django-cors-headers==3.14.0
django-crispy-forms==2.3
django-jazzmin==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.0.0
PyJWT==2.6.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
